activityentry: Disable drop-down completion list
It was causing a lot of refreshes on Wayland, which was causing events
to the window to be dropped. I think the solution is to refactor the
code so that the completion entries in the model are added or removed
individually as necessary, rather than dropping them all and re-adding
them again; but that’s too much refactoring for the time I have at the
moment. So just disable the drop-down and stick with in-line completion.